Fixed IWearWrapper behavior about checking IWear interface status
Now the device is attached correctly and is the run method that will wait. Stopping the device for waiting during the attach step prevents accessing rpc ports of other devices. In this case, it prevented using the XsensMVNControl to launch the calibration and start streaming data. Without this operation, the IWear status would never go to Ok and it resulted in a deadlock state.